requirements.txt
openai>=1.0.0
tqdm
python-dotenv